correction of atom_d.py for 5d Jeff=1/2 systems

+++ src/impurity/atomd/atom_d.py
@@ -199,7 +199,7 @@ class operateLS(object):
         #############################################
             self.Q3d=False
             M2=[]
-            l=(self.Nband-1)/2
+            l=int((self.Nband-1)/2)
             #print 'L is here ', l
             for s in [0.5,-0.5]:
                 for m in range(-l,l+1):
